[WIP] server side image rendering working
