+++ README.md
@@ -65,6 +65,8 @@ Note: that those with an asterisk *(\*)* will have more detailed information , a
     - post a user path recording to a specified trail id.
   - GET `/paths/:pathId` * 
     - retrieves detailed information about a path by a given ID in database. this also will retrieve gpx data.
+  - GET `/paths/:pathId/image/:width/:height`
+    - renders static PNG map on server side given width and height
   - GET `/:trailId/heroPath` * 
     - retrieves detailed information about the canonical path for a given trail data. this also will retrieve gpx data.
   - GET `/:trailId/trailHead` * 
@@ -405,4 +407,23 @@ app.get('/:trailId(\\d+$)*?', (req, res) => {
 
   ```
 
+### React Widget Development
 
+**Server Side Image Rendering**
+
+Moving back to react client development, I first was focused on Recording widget which comprised of all the user submitted paths. One concern was to be rendering all these mapbox widgets for every recording seemed too like it would be overkill / computationally expensive. Instead, similar to actual alltrails, it would make more sense to render a static map (as an image) for each recording on the server, and then show that instead. I managed to do this with 2 npm packages: `@mapbox/mapbox-gl-native`, a node native library which allows for the retrieval of the mapbox tiles and drawing geojson geometry, and than `sharp` for processing the buffers to image output. Then my endpoint  `/paths/:pathId/image/:width/:height` points to a dynamically rendered PNG image! (This tutorial)[https://medium.com/@brendan_ward/creating-a-static-map-renderer-using-the-mapbox-gl-native-nodejs-api-23db560b219e] was very helpful in this endeavor, as well as the npm docs. 
+
+Furthermore, one issue I had was the `map.fitBounds()` was not available for getting the appropriate `zoom` value for the map in this version of mapbox.  this is very userful in client version as it will autozoom to your given lat lng bounds data, but with this I could only supply a number `zoom=0-24`, where 0 is the entire world, and 24 is the most zoomed in. Our trails tended to need to be zoomed anywhere from `8-13`, but I couldn't find a pattern to convert our bounds to this weird arbitrary `zoom` number. Then I found a google groups post that pretty much explained the logic behind google maps' zoom index, analagous to mapbox! With a few tweaks (aka a modifier to shrink / convert to mapbox zoom), voila, i was able to create this number from my bounds and pixel width/height! This can be seen in `staticMap.js`. Snippet.
+
+``` js
+  map.render({
+      width: width,
+      height: height,
+      zoom: staticMap.getBoundsZoomLevel(bounds, {width: width * modifier, height: height * modifier}),  
+      center: [ 
+        (Number(bounds.minlon) + Number(bounds.maxlon)) / 2,
+        (Number(bounds.minlat) + Number(bounds.maxlat)) / 2
+
+      ]
+    }, ...
+```
